Split up pause index.
